requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pandas==2.1.4
lxml==4.9.3

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.0

# Database
sqlalchemy>=2.0.0

# Visualization and analysis
networkx>=3.2.0
matplotlib>=3.8.0
plotly>=5.17.0

# Web interface (for future development)
streamlit>=1.29.0
flask>=3.0.0

# Development tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0 